third_party/blink/web_tests/media/W3C/audio/events/event_order_loadstart_progress-expected.txt

setting src attribute on autoplay video should trigger loadstart then progress event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS successfullyParsed is true

TEST COMPLETE


spec reference
